Tidy up tests and split style tests
