+++ packages/edition-slices/__tests__/ios/edition-slices.test.js
@@ -1,3 +1,3 @@
-import shared from "../shared";
+import shared from "../shared.native";
 
 shared();
