+++ packages/edition-slices/__tests__/web/edition-slices.test.js
@@ -1,3 +1,3 @@
-import shared from "../shared";
+import shared from "../shared.web";
 
 shared();
