+++ packages/edition-slices/fixtures/leadoneandone.js
@@ -1,5 +1,5 @@
 /* eslint-disable sort-keys */
-export default {
+export default ({ imageUrl }) => ({
   __typename: "LeadOneAndOneSlice",
   lead: {
     article: {
@@ -14,10 +14,12 @@ export default {
       leadAsset: {
         crop169: {
           url:
+            imageUrl ||
             "https://www.thetimes.co.uk/imageserver/image/%2Fmethode%2Ftimes%2Fprod%2Fweb%2Fbin%2F8c1333a8-1aa2-11e9-944c-54b267eb465b.jpg?crop=1556%2C875%2C189%2C119"
         },
         crop32: {
           url:
+            imageUrl ||
             "https://www.thetimes.co.uk/imageserver/image/%2Fmethode%2Ftimes%2Fprod%2Fweb%2Fbin%2F8c1333a8-1aa2-11e9-944c-54b267eb465b.jpg?crop=1556%2C1037%2C189%2C38"
         },
         id: "372bc095-34c4-47e4-8b1e-d352f5641ee5",
@@ -77,10 +79,12 @@ export default {
       leadAsset: {
         crop169: {
           url:
+            imageUrl ||
             "https://www.thetimes.co.uk/imageserver/image/%2Fmethode%2Ftimes%2Fprod%2Fweb%2Fbin%2Fde5b8008-1a5b-11e9-944c-54b267eb465b.jpg?crop=2000%2C1125%2C0%2C104"
         },
         crop32: {
           url:
+            imageUrl ||
             "https://www.thetimes.co.uk/imageserver/image/%2Fmethode%2Ftimes%2Fprod%2Fweb%2Fbin%2Fde5b8008-1a5b-11e9-944c-54b267eb465b.jpg?crop=2000%2C1333%2C0%2C0"
         },
         id: "3ab8acf9-ef69-4c06-df64-57521b07d79c",
@@ -115,4 +119,4 @@ export default {
       ]
     }
   }
-};
+});
