+++ packages/edition-slices/package.json
@@ -46,7 +46,9 @@
     "@thetimes/jest-lint": "*",
     "@times-components/eslint-config-thetimes": "0.8.9",
     "@times-components/jest-configurator": "2.1.20",
+    "@times-components/jest-serializer": "3.2.0",
     "@times-components/storybook": "3.2.6",
+    "@times-components/test-utils": "2.1.0",
     "@times-components/webpack-configurator": "2.0.15",
     "babel-cli": "6.26.0",
     "eslint": "5.9.0",
